### 2021-11-23 #3925 @johnveitch
Hi @arfon. Thanks for the invitation to act as a reviewer for JOSS. I had a look through LegWork and it seems to be a very nicely put together package. I would be happy to review it but I've got a pretty full diary before Christmas so I don't think I'll be very swift. Perhaps if we can divide the work between myself and others it would be possible on that timescale, if that would suit you.
