### 2021-11-23 #3925 @cmbiwer
I just finished reviewing another JOSS submission a couple weeks ago, so I'd let others have a chance before me. But if no one else is available, then I can be one of the reviewers.
