### 2021-11-24 #3925 @arfon
:wave: @sbozzolo @dbkeitel – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is: _LEGWORK: A python package for computing the evolution and detectability of stellar-origin gravitational-wave sources with space-based detectors_.
