### 2021-11-24 #3925 @duncanmmacleod
> @duncanmmacleod @cmbiwer @johnveitch – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The submission under consideration is: _LEGWORK: A python package for computing the evolution and detectability of stellar-origin gravitational-wave sources with space-based detectors_.

Sorry @arfon, similar to John I don't think I'll be able to get to this quickly, but could come back to this in the new year if you haven't been able to find a reviewer before then.
