### 2021-11-26 #3925 @dbkeitel
This looks like a very professionally assembled package, and in principle I'd be happy to help with review. However I'm confused by the intended relation between the JOSS paper and the longer one at https://arxiv.org/abs/2111.08717 with the exact same title and very similar abstract. From https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software I see that in principle "co-publication" is fine, but here it looks like the other version also explicitly describes the package and so I'm not sure what the JOSS submission is adding. So before committing to review, maybe @TomWagg could clarify a bit and  @arfon could opine if this seems reasonable.
