### 2021-11-26 #3925 @TomWagg
Hi @dbkeitel, I'm sure @arfon will have more thoughts but my understanding was that this was the point behind co-publication. Our motivation behind also submitting to JOSS is that we are able to get a review of the code itself. Only the manuscript that we submit to AAS will be reviewed and it was important to us that we ensure that not just our writing, but also our package was reviewed and that's what the JOSS submission adds. We think this will help us to improve the code and, should the review be successful, will indicate to others that the package has been written well.

But I'll of course defer to @arfon if we've misunderstand how these co-publications should work!
