### 2021-12-01 #3925 @dbkeitel
Thanks for the clarifications @TomWagg and @arfon , I was not aware of the details of this initiative and from the documentations page I had interpreted the model as more "orthogonal" than this very "parallel" process. Still, thinking as a potential reader / end-user I'd personally find it very confusing to encounter two papers in different journals that overlap so much and even share the title; is that explicitly expected under the agreement with AAS or subject to discussion during review?

Also when you say "a more detailed software review", is this still the original JOSS model that focuses mostly on packaging and documentation as per the check-list, or is an actual in-depth review of the scientific and numerical validity of the code expected? (Similar to what e.g. the LVK collaboration, and probably other large collaborations, consider a "software review".) I'm afraid for the latter I wouldn't have the qualification and time.
