### 2021-12-02 #3925 @arfon
> Still, thinking as a potential reader / end-user I'd personally find it very confusing to encounter two papers in different journals that overlap so much and even share the title; is that explicitly expected under the agreement with AAS or subject to discussion during review?

Yes, I can see how this could be a little confusing. Historically – for the ~25 papers that have been through this process so far – the titles have been identical/very similar between the JOSS and AAS paper in ~25% of cases (like here).

> Also when you say "a more detailed software review", is this still the original JOSS model that focuses mostly on packaging and documentation as per the check-list.

:+1: yes, the JOSS review is the same.

Going back to the origins/purpose of this dual review process for a moment. The AAS paper is likely to be the 'canonical' reference for this software package (although this is of course up to the authors). The AAS paper includes a far more detailed description of what the purpose of the package is, detailed mathematical derivations of the algorithms implemented in the code, detailed example results etc. While the AAS review process is mostly focused on the scientific merit of this software within a specific area of science, the is little/no actual review of the software, its implementation, usability etc. and this is where JOSS steps in.

I guess what I'm saying is that the AAS and JOSS reviews are two halves of a whole review of the work. Due to us not wanting to conflate our editorial processes, when we first established this collaboration we decided it would be most straightforward to simply have two papers go through a normal review, and for them to cite each other.
