### 2021-12-03 #3925 @dbkeitel
Thanks @arfon ! It probably isn't a model I'd be going for myself as an author, but I agree that this submission matches your intended model, and you can sign me up for review.

Also as a purely optional and personal suggestion to @tomwagg and co-authors: to disambiguate titles, the current one seems perfect for JOSS, but maybe you could consider changing over at ApJ to something like `Computing the evolution and detectability of stellar-origin gravitational-wave sources for space-based detectors with the python package LEGWORK`? Then it would be clear that the JOSS paper is a "software reference" while ApJ focuses more on the physics of it. But as @arfon wrote, having the same title seems perfectly fine by editorial agreement between the two journals, so I won't officially complain review-wise if you prefer to keep both the same.
