### 2021-12-05 #3925 @arfon
> Also as a purely optional and personal suggestion to @TomWagg and co-authors: to disambiguate titles, the current one seems perfect for JOSS, but maybe you could consider changing over at ApJ to something like Computing the evolution and detectability of stellar-origin gravitational-wave sources for space-based detectors with the python package LEGWORK? Then it would be clear that the JOSS paper is a "software reference" while ApJ focuses more on the physics of it.

@TomWagg – this isn't required, but I do think this is a good suggestion if possible.
