### 2021-12-06 53e9c5a
separate confusion noises into functions
