### 2021-12-06 3ed14cf
add new Thiele+21 fit
