### 2021-12-06 0f62b56
edit test to include all models of confusion noise
