+++ legwork/tests/test_psd.py
@@ -20,16 +20,18 @@ class Test(unittest.TestCase):
         """check that confusion noise is doing logical things"""
         frequencies = np.logspace(-6, 0, 10000) * u.Hz
 
-        confused = psd.power_spectral_density(frequencies, confusion_noise="robson19")
-        lucid = psd.power_spectral_density(frequencies, confusion_noise=None)
+        models = ["robson19", "huang20", "thiele21"]
+        instruments = ["LISA", "TianQin", "LISA"]
+        for model, instrument in zip(models, instruments):
+            confused = psd.get_confusion_noise(frequencies, model=model)
+            lucid = psd.get_confusion_noise(frequencies, model=None)
 
-        # ensure confusion noise only adds to noise
-        self.assertTrue(np.all(confused >= lucid))
+            # ensure confusion noise only adds to noise
+            self.assertTrue(np.all(confused >= lucid))
 
-        # ensure that it doesn't affect things at low or high frequency
-        safe = np.logical_or(frequencies < 1e-4 * u.Hz,
-                             frequencies > 1e-2 * u.Hz)
-        self.assertTrue(np.allclose(confused[safe], lucid[safe]))
+            # ensure that it doesn't affect things at low or high frequency
+            safe = np.logical_or(frequencies < 1e-4 * u.Hz, frequencies > 1e-2 * u.Hz)
+            self.assertTrue(np.allclose(confused[safe], lucid[safe]))
 
     def test_mission_length_effect(self):
         """check that increasing the mission length isn't changing
