### 2021-12-06 2db05c3
fix mixed units in visualisation
