### 2021-12-06 #3925 @TomWagg
Agreed @arfon, thanks for the feedback @dbkeitel! We've updated the title of the ApJ paper to `Computing the evolution and detectability of stellar-origin gravitational-wave sources in space-based detectors using the LEGWORK python package` (as you can see [here](https://raw.githubusercontent.com/TeamLEGWORK/LEGWORK-paper/main-pdf/ms.pdf))

I think perhaps this could be something to add to the guidance on joint JOSS/AAS submissions if there is a preference for having distinct titles between the two?
