### 2021-12-08 #3925 @TomWagg
Actually, having discussed this with my co-authors, they suggested that perhaps it isn't sensible to change the title of the ApJ paper. Since, changing the title doesn't really change the _content_ of the paper, and so it could become even more confusing if two papers with extremely similar content have different titles. Whereas having the same title can make it extremely clear that the two papers are a joint submission.

But if you still feel strongly that they should have different names we shall of course follow your guidance!
