### 2021-12-17 #3998 @whedon
[REVIEW]: LEGWORK: A python package for computing the evolution and detectability of stellar-origin gravitational-wave sources with space-based detectors
**Submitting author:** @TomWagg (<a href="http://orcid.org/0000-0001-6147-5761">Thomas Wagg</a>)
**Repository:** <a href="https://github.com/TeamLEGWORK/LEGWORK" target ="_blank">https://github.com/TeamLEGWORK/LEGWORK</a>
**Version:** v0.1.3
