### 2021-12-17 #3998 @arfon
We aim for the review process to be completed within about 4-6 weeks but given we're entering the holiday season, I suspect things might move slower than this.
