### 2022-01-04 TeamLEGWORK/LEGWORK#60 @dbkeitel
license question
**Which documentation page is causing an issue?**
[LICENSE.rst](https://github.com/TeamLEGWORK/LEGWORK/blob/main/LICENSE.rst)

**What is the problem?**
Hi @katiebreivik @TomWagg getting started with the JOSS review ( -> https://github.com/openjournals/joss-reviews/issues/3998 ), I noticed that your license file only lists Katie as the `(c)` holder, while the code report gives roughly similarl contributions from the two of you and the paper submission has 3 authors (also Selma). Maybe it should be updated?

I know very little about licensing and copyright law though, so this is a purely optional suggestion.
