### 2022-01-04 TeamLEGWORK/LEGWORK#61 @dbkeitel
list dependencies in README?
**Which documentation page is causing an issue?**
README.md

**What is the problem?**
From JOSS review ( -> openjournals/joss-reviews#3998 ):
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

**Suggested solution (optional)**

In a sense, the installation instructions already list the dependencies via the suggested conda command. There's also a requirements.txt file, so I'm going to tick the box over on JOSS already. However, it would be nice for the user to also explicitly list the dependencies in the README, so that they also show up easily visible over on PyPI.
