### 2022-01-04 #3998 @dbkeitel
Getting started on the checklist, a first set of notes for the early items (basically the metadata):

> License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

Yes. See https://github.com/TeamLEGWORK/LEGWORK/issues/60 for a quick check on which authors it should list.

> Contribution and authorship: Has the submitting author (@TomWagg) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

According to whedon's analytics posted above, @TomWagg and @katiebreivik have roughly similar contributions to the source code. The third author of the paper, Selma de Mink, does not appear to have contributed directly to the repository. I don't doubt her involvement in the scientific development of course, but just to have a full record, it would be nice to have a short note of confirmation here.

> Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

Very evidently yes, as demonstrated by the companion ApJ paper.

> Installation: Does installation proceed as outlined in the documentation?

Installing both into a `venv` with `pip` and into a `conda` env work as expected. It would be nice to also package for `conda-forge` directly, but the documentation explains that this is currently not possible due to one external dependency not living in that system yet.

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Both README and the main docs have a handful of concise but very clear intro sentences that fulfill this role well. The tutorials then cover the actual science scope in much more detail, which is a nice way to do so without putting a wall of text right at the start.

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The instructions are clear and work. See https://github.com/TeamLEGWORK/LEGWORK/issues/61 for listing dependencies more visibly.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

There's no detailed contributor's guide on coding style and details like that, but right in the main README (and hence PyPI page) there are links to the issue tracker and an encouragement to submit pull requests. The readthedocs page also contains a prominent "Submit an issue" link. I think this is sufficient.
