### 2022-01-04 #3998 @dbkeitel
On the JOSS manuscript itself: overall I think it's in great shape, so I've checked off the first 4 items already. Just the references need a little extra work.

Main comments:
- [ ] As discussed in pre-review, the JOSS manuscript should be updated to explicitly point the reader to the ApJ companion paper for more details.
- [ ] The current reference list seems long to me by JOSS standards, with the majority being general references for sources not covered by the package itself (L28-32) and for recent papers on the "stellar binaries" that the package _does_ cover (L40-43). I haven't checked in detail, but presumably these are all cited from the ApJ paper too, so an option would be to condense the lists here and refer to ApJ "for further references". However, if @arfon would say that the list is not excessive for JOSS standards, I'm also ok with it staying as is. (Note I suggest a few additions below on other topics.)
- [ ] Since this package covers not only orbital evolutions but also strains, what do the authors think about saying something about the relationship to more waveform-focused packages? Coming from the ground-based side of the field, I'd say a reference to LALSuite with its various Taylor WF implementations might be appropriate ("might" because I never tried to generate anything on LISA timescales for it, and it may well be completely unsuitable for that, so I leave the call to the authors), and I'd expect there must be some similar LISA-dedicated packages I'm not yet familiar with...?

And a few more quick items:
- [ ] L26: should cite a LISA mission/detector ref
- [ ] L28 and several times later: fix formatting for "e.g." lists of citations (remove parentheses, add commas and spaces)
- [ ] L38: should cite LIGO first detection paper
- [ ] L61: should cite a TianQin design paper
- [ ] L64: there are no "derivations described below"; refer to ApJ instead?
- [ ] L68: not clear what the "e.g." is doing in the middle of the references here

I may also have additional comments on the manuscript later, after having had a closer look at the implementation and documentation.
