### 2022-01-04 TeamLEGWORK/LEGWORK#60 @TomWagg
Good point, I think the license should contain any authors who directly contributed to the code so I'll add myself to the LICENSE.rst file. (Selma helped with the scientific development/suggested features/helped with the manuscript writing but didn't contribute code directly)
