### 2022-01-04 TeamLEGWORK/LEGWORK#61 @TomWagg
Agreed! I'll add them to the README now.
