### 2022-01-04 #3998 @TomWagg
Thanks for the comments @dbkeitel! First to address the metadata comments:

I have editted the LICENSE.rst to include all authors who contributed code and added the dependencies explicitly to the README from your comments (see [PR #62](https://github.com/TeamLEGWORK/LEGWORK/pull/62)).

> The third author of the paper, Selma de Mink, does not appear to have contributed directly to the repository. I don't doubt her involvement in the scientific development of course, but just to have a full record, it would be nice to have a short note of confirmation here.

As you say, Selma has not contributed any code directly. She has however helped with the development of LEGWORK indirectly. Selma helped me with the derivation of the equations used in LEGWORK when I was first working on implementing them for simulations. Her main contributions have come from suggestions of features included in LEGWORK and about the structure of the package. She also helped significantly with the preparation of the manuscript.
