### 2022-01-04 3e7334e
add LISA detector citation
