+++ paper.bib
@@ -1,3 +1,20 @@
+@ARTICLE{Amaro-Seoane+2017,
+       author = {{Amaro-Seoane}, Pau and {Audley}, Heather and {Babak}, Stanislav and {Baker}, John and {Barausse}, Enrico and {Bender}, Peter and {Berti}, Emanuele and {Binetruy}, Pierre and {Born}, Michael and {Bortoluzzi}, Daniele and {Camp}, Jordan and {Caprini}, Chiara and {Cardoso}, Vitor and {Colpi}, Monica and {Conklin}, John and {Cornish}, Neil and {Cutler}, Curt and {Danzmann}, Karsten and {Dolesi}, Rita and {Ferraioli}, Luigi and {Ferroni}, Valerio and {Fitzsimons}, Ewan and {Gair}, Jonathan and {Gesa Bote}, Lluis and {Giardini}, Domenico and {Gibert}, Ferran and {Grimani}, Catia and {Halloin}, Hubert and {Heinzel}, Gerhard and {Hertog}, Thomas and {Hewitson}, Martin and {Holley-Bockelmann}, Kelly and {Hollington}, Daniel and {Hueller}, Mauro and {Inchauspe}, Henri and {Jetzer}, Philippe and {Karnesis}, Nikos and {Killow}, Christian and {Klein}, Antoine and {Klipstein}, Bill and {Korsakova}, Natalia and {Larson}, Shane L and {Livas}, Jeffrey and {Lloro}, Ivan and {Man}, Nary and {Mance}, Davor and {Martino}, Joseph and {Mateos}, Ignacio and {McKenzie}, Kirk and {McWilliams}, Sean T and {Miller}, Cole and {Mueller}, Guido and {Nardini}, Germano and {Nelemans}, Gijs and {Nofrarias}, Miquel and {Petiteau}, Antoine and {Pivato}, Paolo and {Plagnol}, Eric and {Porter}, Ed and {Reiche}, Jens and {Robertson}, David and {Robertson}, Norna and {Rossi}, Elena and {Russano}, Giuliana and {Schutz}, Bernard and {Sesana}, Alberto and {Shoemaker}, David and {Slutsky}, Jacob and {Sopuerta}, Carlos F. and {Sumner}, Tim and {Tamanini}, Nicola and {Thorpe}, Ira and {Troebs}, Michael and {Vallisneri}, Michele and {Vecchio}, Alberto and {Vetrugno}, Daniele and {Vitale}, Stefano and {Volonteri}, Marta and {Wanner}, Gudrun and {Ward}, Harry and {Wass}, Peter and {Weber}, William and {Ziemer}, John and {Zweifel}, Peter},
+        title = "{Laser Interferometer Space Antenna}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2017,
+        month = feb,
+          eid = {arXiv:1702.00786},
+        pages = {arXiv:1702.00786},
+archivePrefix = {arXiv},
+       eprint = {1702.00786},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2017arXiv170200786A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @ARTICLE{Begelman+1980,
        author = {{Begelman}, M.~C. and {Blandford}, R.~D. and {Rees}, M.~J.},
         title = "{Massive black hole binaries in active galactic nuclei}",
