+++ paper.md
@@ -46,15 +46,15 @@ bibliography: paper.bib
 
 # Statement of need
 
-The planned space-based gravitational wave detector LISA (Laser Interferometer Space Antenna, @Amaro-Seoane+2017) will present an entirely new view of gravitational waves by focusing on lower frequencies ($10^{-5} < f / \unit{Hz} < 10^{-1}$) than ground-based detectors. This will enable the study of many new source classes including mergers of supermassive black holes e.g.[@Begelman+1980; @Klein+2016; @Bellovary2019], extreme mass ratio inspirals e.g.[@Berti2006; @Barack2007; @Babak2017; @Moore2017], and cosmological GW backgrounds e.g.[@Bartolo2016; @Caprini2016; @Caldwell2019]. However, this frequency regime is also of interest for the detection of local stellar-mass binaries during their inspiral phase. LISA is expected to detect Galactic stellar-origin binaries containing combinations of white dwarfs, neutron stars, and black holes, ranging from the numerous double white dwarf population, to the rare but loud double black hole population.
+The planned space-based gravitational wave detector LISA (Laser Interferometer Space Antenna, @Amaro-Seoane+2017) will present an entirely new view of gravitational waves by focusing on lower frequencies ($10^{-5} < f / \unit{Hz} < 10^{-1}$) than ground-based detectors. This will enable the study of many new source classes including mergers of supermassive black holes [e.g. @Begelman+1980; @Klein+2016; @Bellovary2019], extreme mass ratio inspirals [e.g. @Berti2006; @Barack2007; @Babak2017; @Moore2017], and cosmological GW backgrounds [e.g. @Bartolo2016; @Caprini2016; @Caldwell2019]. However, this frequency regime is also of interest for the detection of local stellar-mass binaries during their inspiral phase. LISA is expected to detect Galactic stellar-origin binaries containing combinations of white dwarfs, neutron stars, and black holes, ranging from the numerous double white dwarf population, to the rare but loud double black hole population.
 
-The potential to detect stellar-origin sources with LISA has been studied in the past e.g.[@Nelemans+2001; @Liu+2009; @Liu+2014; @Ruiter+2010; @Belczynski+2010; @Nissanke+2012]. More recently, the direct detection of gravitational waves with ground-based detectors has led to renewed interest in this topic with many recent papers addressing the issue e.g.[@Christian+2017; @Kremer+2017; @Kremer+2018; @Korol+2017; @Korol+2018; @Korol+2019; @Korol+2020; @Lamberts+2018; @Lamberts+2019; @Fang+2019; @Andrews+2020; @Lau+2020; @Breivik+2020; @Breivik+2020a; @Roebber+2020; @Chen+2020; @Sesana+2020; @Shao+2021].
+The potential to detect stellar-origin sources with LISA has been studied in the past [e.g. @Nelemans+2001; @Liu+2009; @Liu+2014; @Ruiter+2010; @Belczynski+2010; @Nissanke+2012]. More recently, the direct detection of gravitational waves with ground-based detectors has led to renewed interest in this topic with many recent papers addressing the issue [e.g. @Christian+2017; @Kremer+2017; @Kremer+2018; @Korol+2017; @Korol+2018; @Korol+2019; @Korol+2020; @Lamberts+2018; @Lamberts+2019; @Fang+2019; @Andrews+2020; @Lau+2020; @Breivik+2020; @Breivik+2020a; @Roebber+2020; @Chen+2020; @Sesana+2020; @Shao+2021].
 
 Each of these studies require making estimates of the signal-to-noise ratio of individual binary systems and possibly the slow gravitational wave inspiral that lead to the present-day parameters. So far, most studies made use of custom made codes which have not been made publicly available. 
 
[…]
