+++ paper.md
@@ -54,7 +54,7 @@ Each of these studies require making estimates of the signal-to-noise ratio of i
 
 We believe that the large renewed interest in LISA and the stellar-origin sources it may detect will lead to many more studies in the near future that would need similar computations. This leads to a significant amount of redundancy which, at best results in extra work for each individual and at worst leads to an increased chance of introducing mistakes and inconsistencies when translating the necessary expressions to software.
 
-\lw{} is a Python package designed to streamline the process of making predictions of LISA detection rates for stellar-origin binaries such that it is as fast, reliable and simple as possible. This goal makes \lw{} unique among other implementations of gravitational-wave tools in the literature, which focus on a more broad coverage of the gravitational-wave spectrum and source classes, rather than an optimised approach for certain sources [e.g. @Moore+2015; @Yi2021]. With \lw{} one can evolve the orbits of a binary or a collection of binaries and calculate their strain amplitudes for any range of frequency harmonics. One can compute the sensitivity curve for LISA or other future gravitational wave detectors (e.g. TianQin's curve, or that of a custom instrument) and use it to compute the signal-to-noise ratio of a collection of sources. Furthermore, \lw{} provides tools to visualise all of the results with easy-to-use plotting functions. Finally, \lw{} is fully tested to check for consistency in the derivations described below.
+\lw{} is a Python package designed to streamline the process of making predictions of LISA detection rates for stellar-origin binaries such that it is as fast, reliable and simple as possible. This goal makes \lw{} unique among other implementations of gravitational-wave tools in the literature, which focus on a more broad coverage of the gravitational-wave spectrum and source classes, rather than an optimised approach for certain sources [e.g. @Moore+2015; @Yi2021]. With \lw{} one can evolve the orbits of a binary or a collection of binaries and calculate their strain amplitudes for any range of frequency harmonics. One can compute the sensitivity curve for LISA or other future gravitational wave detectors (e.g. TianQin, @Luo+2016, or that of a custom instrument) and use it to compute the signal-to-noise ratio of a collection of sources. Furthermore, \lw{} provides tools to visualise all of the results with easy-to-use plotting functions. Finally, \lw{} is fully tested to check for consistency in the derivations described below.
 
 Specifically, we implement the expressions by @Peters+1963 and @Peters+1964 for the evolution of binary orbits due to the emission of gravitational waves, equations for the strain amplitudes and signal-to-noise ratios of binaries from various papers [e.g. @Flanagan+1998; @Finn+2000; @Cornish2003; @Barack+2004; @Moore+2015] and approximations for the LISA and TianQin sensitivity curves given in @Robson+2019 and @Huang+2020 respectively.
 
