+++ paper.bib
@@ -692,6 +692,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Luo+2016,
+       author = {{Luo}, Jun and {Chen}, Li-Sheng and {Duan}, Hui-Zong and {Gong}, Yun-Gui and {Hu}, Shoucun and {Ji}, Jianghui and {Liu}, Qi and {Mei}, Jianwei and {Milyukov}, Vadim and {Sazhin}, Mikhail and {Shao}, Cheng-Gang and {Toth}, Viktor T. and {Tu}, Hai-Bo and {Wang}, Yamin and {Wang}, Yan and {Yeh}, Hsien-Chi and {Zhan}, Ming-Sheng and {Zhang}, Yonghe and {Zharov}, Vladimir and {Zhou}, Ze-Bing},
+        title = "{TianQin: a space-borne gravitational wave detector}",
+      journal = {Classical and Quantum Gravity},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, General Relativity and Quantum Cosmology},
+         year = 2016,
+        month = feb,
+       volume = {33},
+       number = {3},
+          eid = {035010},
+        pages = {035010},
+          doi = {10.1088/0264-9381/33/3/035010},
+archivePrefix = {arXiv},
+       eprint = {1512.02076},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016CQGra..33c5010L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Peters+1963,
        author = {{Peters}, P.~C. and {Mathews}, J.},
         title = "{Gravitational Radiation from Point Masses in a Keplerian Orbit}",
