+++ paper.md
@@ -56,7 +56,7 @@ We believe that the large renewed interest in LISA and the stellar-origin source
 
 \lw{} is a Python package designed to streamline the process of making predictions of LISA detection rates for stellar-origin binaries such that it is as fast, reliable and simple as possible. This goal makes \lw{} unique among other implementations of gravitational-wave tools in the literature, which focus on a more broad coverage of the gravitational-wave spectrum and source classes, rather than an optimised approach for certain sources [e.g. @Moore+2015; @Yi2021]. With \lw{} one can evolve the orbits of a binary or a collection of binaries and calculate their strain amplitudes for any range of frequency harmonics. One can compute the sensitivity curve for LISA or other future gravitational wave detectors (e.g. TianQin, @Luo+2016, or that of a custom instrument) and use it to compute the signal-to-noise ratio of a collection of sources. Furthermore, \lw{} provides tools to visualise all of the results with easy-to-use plotting functions. Finally, \lw{} is fully tested to check for consistency in the derivations described below.
 
-Specifically, we implement the expressions by @Peters+1963 and @Peters+1964 for the evolution of binary orbits due to the emission of gravitational waves, equations for the strain amplitudes and signal-to-noise ratios of binaries from various papers [e.g. @Flanagan+1998; @Finn+2000; @Cornish2003; @Barack+2004; @Moore+2015] and approximations for the LISA and TianQin sensitivity curves given in @Robson+2019 and @Huang+2020 respectively.
+Specifically, we implement the expressions by @Peters+1963 and @Peters+1964 for the evolution of binary orbits due to the emission of gravitational waves, equations for the strain amplitudes and signal-to-noise ratios of binaries from various papers [e.g., @Flanagan+1998; @Finn+2000; @Cornish2003; @Barack+2004; @Moore+2015] and approximations for the LISA and TianQin sensitivity curves given in @Robson+2019 and @Huang+2020 respectively.
 
 The open-source nature of the project means that new users as well as seasoned experts in the field can work together in a collaborative setting to consider new features and enhancements to the package as well as check the implementation. At the same time, with our thorough online documentation, derivations and tutorials, we hope \lw{} can make this functionality more accessible to the broader scientific community.
 
