### 2022-01-04 19bf72d
add apj paper citation
