### 2022-01-04 #3998 @TomWagg
Now the manuscript comments:
> - [x] As discussed in pre-review, the JOSS manuscript should be updated to explicitly point the reader to the ApJ companion paper for more details.

I wasn't sure exactly how to phrase this but I wrote "This paper is jointly submitted to ApJ, see Wagg+2021.", please let me know if there is a different way I should write this.

**References**: I'll wait for @arfon's thoughts on the length of references before touching this

**Waveform packages**: I'm honestly very unfamiliar with waveform things and so don't know a LISA waveform package off the top of my head. Maybe something like [lisacattools](https://github.com/tlittenberg/lisacattools) could do what you have in mind? I imagine @katiebreivik will know more about this than me and can comment better?

I've addressed most of the checkboxes in the latest commits to the `joss` branch, thanks for suggesting these improvements!

The one thing I am having trouble with is the formatting of (e.g. \<list of citations\>) type citations. Could @arfon perhaps help me with the syntax? Based on [this page](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), a list of citations should look like this `[@author1:2001; @author2:2001]`. But if I then do `[e.g. @author1:2001; @author2:2001]`, the e.g. sometimes appears between a random set of authors rather than at the start (this is causing the problem on L68 that @dbkeitel highlighted). What would be the proper way to do this?
