### 2022-01-04 #3998 @dbkeitel
Thanks for the changes @TomWagg - I'll get back to the rest of the review sometime soon.

Regarding reference formatting, I think you can just use plain `@` syntax without the square brackets and format the list by hand as you wish - e.g. in my only JOSS paper so far, I had a block
```
Many search methods have been developed and applied on LIGO [@TheLIGOScientific:2014jea] and Virgo [@TheVirgo:2014hva] data. See @Prix:2009oha, @Riles:2017evm, and @Sieniawska:2019hmd for reviews of the field.
```
which came out formatted like this:
![image](https://user-images.githubusercontent.com/28396587/148092192-99087ce8-15e3-49d9-93f3-faed41a2c4bf.png)
