+++ paper.bib
@@ -14,7 +14,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Amaro-Seoane+2017,
+@ARTICLE{AmaroSeoane+2017,
        author = {{Amaro-Seoane}, Pau and {Audley}, Heather and {Babak}, Stanislav and {Baker}, John and {Barausse}, Enrico and {Bender}, Peter and {Berti}, Emanuele and {Binetruy}, Pierre and {Born}, Michael and {Bortoluzzi}, Daniele and {Camp}, Jordan and {Caprini}, Chiara and {Cardoso}, Vitor and {Colpi}, Monica and {Conklin}, John and {Cornish}, Neil and {Cutler}, Curt and {Danzmann}, Karsten and {Dolesi}, Rita and {Ferraioli}, Luigi and {Ferroni}, Valerio and {Fitzsimons}, Ewan and {Gair}, Jonathan and {Gesa Bote}, Lluis and {Giardini}, Domenico and {Gibert}, Ferran and {Grimani}, Catia and {Halloin}, Hubert and {Heinzel}, Gerhard and {Hertog}, Thomas and {Hewitson}, Martin and {Holley-Bockelmann}, Kelly and {Hollington}, Daniel and {Hueller}, Mauro and {Inchauspe}, Henri and {Jetzer}, Philippe and {Karnesis}, Nikos and {Killow}, Christian and {Klein}, Antoine and {Klipstein}, Bill and {Korsakova}, Natalia and {Larson}, Shane L and {Livas}, Jeffrey and {Lloro}, Ivan and {Man}, Nary and {Mance}, Davor and {Martino}, Joseph and {Mateos}, Ignacio and {McKenzie}, Kirk and {McWilliams}, Sean T and {Miller}, Cole and {Mueller}, Guido and {Nardini}, Germano and {Nelemans}, Gijs and {Nofrarias}, Miquel and {Petiteau}, Antoine and {Pivato}, Paolo and {Plagnol}, Eric and {Porter}, Ed and {Reiche}, Jens and {Robertson}, David and {Robertson}, Norna and {Rossi}, Elena and {Russano}, Giuliana and {Schutz}, Bernard and {Sesana}, Alberto and {Shoemaker}, David and {Slutsky}, Jacob and {Sopuerta}, Carlos F. and {Sumner}, Tim and {Tamanini}, Nicola and {Thorpe}, Ira and {Troebs}, Michael and {Vallisneri}, Michele and {Vecchio}, Alberto and {Vetrugno}, Daniele and {Vitale}, Stefano and {Volonteri}, Marta and {Wanner}, Gudrun and {Ward}, Harry and {Wass}, Peter and {Weber}, William and {Ziemer}, John and {Zweifel}, Peter},
         title = "{Laser Interferometer Space Antenna}",
       journal = {arXiv e-prints},
@@ -65,7 +65,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Bellovary2019,
+@ARTICLE{Bellovary+2019,
        author = {{Bellovary}, Jillian M. and {Cleary}, Colleen E. and {Munshi}, Ferah and {Tremmel}, Michael and {Christensen}, Charlotte R. and {Brooks}, Alyson and {Quinn}, Thomas R.},
         title = "{Multimessenger signatures of massive black holes in dwarf galaxies}",
       journal = {\mnras},
@@ -83,7 +83,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Berti2006,
+@ARTICLE{Berti+2006,
[…]
