### 2022-01-04 #3998 @TomWagg
Ah gotcha, thanks @dbkeitel, I hadn't considered that I could just put them in parentheses manually! I've done this, though now there are nested parentheses if that is okay? I could also just remove the "e.g." but that doesn't seem reasonable as these are just examples of papers that address those topics.
