### 2022-01-04 #3998 @TomWagg
I've actually just realised why the "e.g." appears at a weird location. I think JOSS uses pandoc cite-proc which sorts the references by author and then by date. So the "e.g". always goes with whichever citation I put first in the paper.md file...but they then get sorted and so the "e.g." moves with the citation if it doesn't come first alphabetically. I thought this would mean that I can fix the stray "e.g." by just putting them next to the citation that comes first alphabetically but this then _affects_ the sorting! btw @arfon, whilst looking for examples I saw [this recent JOSS paper](https://joss.theoj.org/papers/10.21105/joss.03887) and I think this problem of strangely placed "e.g." also affects it (see the citation list in the summary).

Is there any way to change the sorting that JOSS uses for list of citations? I list of all the citations chronologically in the paper rather than by author name and it would be nice to keep them like that.
