### 2022-01-05 TeamLEGWORK/LEGWORK#63 @dbkeitel
Zenodo releases
**Describe the new feature you'd like**
This is not required for JOSS review, just a suggestion for enhancing best practices by your downstream users: you could consider archiving releases at Zenodo, which assigns a DOI to each version, and then recommending on the "Citing LEGWORK" page that users cite the specific release version they used in addition to the general journal reference.

**Any suggestions for how to implement this**
As you already use github's release feature, this is quite easy, as Zenodo only needs to be set up once and then automatically "pulls" the release once it is made on github. Author metadata is easiest to manage with an additional zenodo.json in the github repo, though.

---
