+++ changelog.md
@@ -73,3 +73,7 @@ Small changes to visualisation code and updates to tutorials/demos with the new 
 ## 0.1.6
 *TW 18/11/21*
 - Add a factor of 10/3 to the TianQin sensitivity curve to make it consistent with LISA (thanks to Yi-Ming Hu for pointing this out!)
+
+## 0.1.7
+*TW 05/01/21*
+- Link GitHub releases to Zenodo
