### 2022-01-05 8108e23
make reqs match in setup.cfg and requirements.txt
