+++ setup.cfg
@@ -13,10 +13,10 @@ long_description_content_type = text/markdown
 python_requires = >=3.7
 packages = find: 
 install_requires = 
-    numpy >= 1.16'
-    scipy >= 1.5.0'
-    astropy >= 4.0'
     numba >= 0.50'
+    numpy >= 1.16,< 1.20'
+    astropy >= 4.0'
+    scipy >= 1.5.0'
     matplotlib >= 3.3.2'
     seaborn >= 0.11.1'
     schwimmbad >= 0.3.2'
