### 2022-01-05 c79aa48
downgrade numba slightly
