+++ setup.cfg
@@ -13,7 +13,7 @@ long_description_content_type = text/markdown
 python_requires = >=3.7
 packages = find: 
 install_requires = 
-    numba >= 0.50'
+    numba >= 0.50,< 0.54'
     numpy >= 1.16,< 1.20'
     astropy >= 4.0'
     scipy >= 1.5.0'
