### 2022-01-05 5888ed6
clarify harmonics required function in comments
