### 2022-01-05 409c2f2
allow manual execution of pypi_upload for testing
