+++ .github/workflows/pypi_upload.yml
@@ -6,6 +6,7 @@ name: Upload Python Package to PyPi
 on:
   release:
     types: [created]
+  workflow_dispatch:
 
 jobs:
   deploy:
