+++ .github/workflows/pypi_upload.yml
@@ -11,11 +11,13 @@ on:
 jobs:
   deploy:
     runs-on: ubuntu-latest
-    python-version: 3.9
 
     steps:
     - uses: actions/checkout@v2
     - uses: actions/setup-python@v2
+    - uses: actions/setup-python@v2
+      with:
+        python-version: '3.9'
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
