### 2022-01-05 7232b85
add test for bad confusion noise input
