### 2022-01-05 babb246
fix comments in psd to include confusion noise
