### 2022-01-05 e51a5fa
account for different convention in Huang+20 cn
