### 2022-01-05 96a394d
default confusion noise auto selects based on instrument
