### 2022-01-05 a228ced
auto select arm length based on instrument
