+++ legwork/tests/test_sources.py
@@ -436,8 +436,9 @@ class Test(unittest.TestCase):
         """ ensuring that updating the sc params always works """
         original_sc_params = {
             "instrument": "LISA",
+            "custom_psd": None,
             "t_obs": 4 * u.yr,
-            "L": 2.5e9 * u.m,
+            "L": "auto",
             "approximate_R": False,
             "confusion_noise": "auto"
         }
@@ -445,12 +446,12 @@ class Test(unittest.TestCase):
         sources = source.Source(m_1=1 * u.Msun, m_2=1 * u.Msun, f_orb=1e-3 * u.Hz, ecc=0.2, dist=10*u.kpc,
                                 sc_params=original_sc_params)
 
-        sources.update_sc_params({"instrument": "TianQin", "L": np.sqrt(3) * 1e5 * u.km})
+        sources.update_sc_params({"instrument": "TianQin"})
 
         correct_final_sc_params = {
             "instrument": "TianQin",
             "t_obs": 4 * u.yr,
-            "L": np.sqrt(3) * 1e5 * u.km,
+            "L": "auto",
             "approximate_R": False,
             "confusion_noise": "auto",
             "custom_psd": None,
