### 2022-01-05 9322aa8
fix test to include confusion noise param
