### 2022-01-05 8dcf770
auto select mission length based on instrument
