### 2022-01-05 9c3eca9
test custom confusion noise in TianQin
