+++ .github/workflows/check_reminders.yml
@@ -1,0 +1,17 @@
+# check reminders and if one is ready then issue it!
+# the schedule uses this format btw https://pubs.opengroup.org/onlinepubs/9699919799/utilities/crontab.html#tag_20_25_07
+# so I've set it to check every minute (I think)
+
+name: 'check reminders'
+
+on:
+  schedule:
+    - cron: '0 * * * *'
+
+jobs:
+  reminder:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: check reminders and notify
+        uses: agrc/reminder-action@v1
