### 2022-01-06 TeamLEGWORK/LEGWORK#75 @dbkeitel
UnitConversionError in plot_sensitivity_curve()
**Describe the bug**
Following the visualisation tutorial, the `plot_sensitivity_curve` function doesn't work for me:

**Example for reproducing**

```
import legwork.visualisation as vis
fig, ax = vis.plot_sensitivity_curve()
Traceback (most recent call last):
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/quantity.py", line 893, in to_value
    scale = self.unit._to(unit)
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/core.py", line 1098, in _to
    raise UnitConversionError(
astropy.units.core.UnitConversionError: 'Unit("Hz")' is not a scaled version of 'Unit(dimensionless)'

Traceback (most recent call last):
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/quantity.py", line 1263, in __float__
    return float(self.to_value(dimensionless_unscaled))
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/quantity.py", line 896, in to_value
    value = self._to_value(unit, equivalencies)
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/quantity.py", line 803, in _to_value
    return self.unit.to(unit, self.view(np.ndarray),
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/core.py", line 1135, in to
    return self._get_converter(Unit(other),
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/core.py", line 1066, in _get_converter
    raise exc
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/core.py", line 1051, in _get_converter
    return self._apply_equivalencies(
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/core.py", line 1027, in _apply_equivalencies
    raise UnitConversionError(
astropy.units.core.UnitConversionError: 'Hz' (frequency) and '' (dimensionless) are not convertible

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/legwork/visualisation.py", line 314, in plot_sensitivity_curve
    ax.fill_between(frequency_range, 0, noise_amplitude, alpha=alpha, color=color)
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/matplotlib/__init__.py", line 1412, in inner
    return func(ax, *map(sanitize_sequence, args), **kwargs)
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/matplotlib/axes/_axes.py", line 5252, in fill_between
    return self._fill_between_x_or_y(
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/matplotlib/axes/_axes.py", line 5225, in _fill_between_x_or_y
    pts[0] = start
  File "/home/dkeitel/bin/venvs/legwork/lib/python3.9/site-packages/astropy/units/quantity.py", line 1265, in __float__
    raise TypeError('only dimensionless scalar quantities can be '
TypeError: only dimensionless scalar quantities can be converted to Python scalars
```

 - OS: Debian 11
 - Version: 0.1.6
 - installed in a venv
 - astropy 5.0
 - Run from: interactive python 3.9.2 session
 - same for a conda env with python 3.9.7
