### 2022-01-06 TeamLEGWORK/LEGWORK#75 @dbkeitel
Interestingly, this code block from later in the tutorial does work for me:
```
# compare LISA and TianQin
fig, ax = vis.plot_sensitivity_curve(show=False, fill=False, instrument="LISA", label="LISA")
fig, ax = vis.plot_sensitivity_curve(show=False, fill=False, instrument="TianQin", label="TianQin",
                                     color=None, fig=fig, ax=ax)
ax.legend()
plt.show()
```
Seems like somethings' wrong with the defaults.
