### 2022-01-06 TeamLEGWORK/LEGWORK#75 @dbkeitel
...but if I do `fill=True` in that second case, it fails again.
