### 2022-01-06 #3998 @dbkeitel
Review update: I'm mostly through with digging as deep into the documentation and code as I think is required and am overall very impressed with the state of the package.

- I've opened a few issues on minor usability and documentation aspects.
- There's a few physics points where I'm a bit confused right now about the choices made, and I've left the "functionality" checkbox open for the moment to remind me of those; but I'll have to properly read the full ApJ manuscript first before I get back to asking about those, and they're likely irrelevant for JOSS scope.
- I've **not** checked the correctness of the equations as documented, **nor** that the code matches the written equations exactly, **nor** the numerical accuracy of anything, beyond the quite impressive level of checks already accessible through the tutorials and test suite. As discussed in pre-review, my understanding is that a more detailed "correctness" review is not required for JOSS.
- I've gone through the basic "tutorials" in detail and verified that they run as intended (except for one visualisation issue -> https://github.com/TeamLEGWORK/LEGWORK/issues/75 ), but only skimmed through the more advanced "demos".
- The documentation and test suite are absolutely **exemplary** :clap: .
  -  Congratulations on 100% coverage (and those tests I've spot-checked not only call the functions and check they don't crash, but actually check outputs as well :+1: ).
  - Maybe the only high-level weakness of the documentation is that users who think "tutorials are for newbies" and skip straight to the API documentation will miss various relevant comments that are _only_ included in the tutorials; but there are nice boxes (with puns) pointing back from the modules to the tutorials, and writing redundant documentation is never fun, so I think that's fine as is.
  - The tutorials themselves are very clear, easy to apply, and include those already-mentioned extra tips and background details.
  - API documentation is a bit sparse on the level of the main docstrings for each method/function, but the input/output documentations make up for it with a very good level of detail.
-  Another comment, mostly thinking towards future maintainability, is that there are a bunch of things hardcoded to particular models from particular papers (e.g. LISA instrumental and confusion noise as well as the set of verification binaries), which may need to change in the future. However it seems to me that the architecture of the package is laid out well enough that this should be possible even in a backwards-compatible way, by e.g. introducing new string arguments for the names of models into those functions.
