### 2022-01-06 #3998 @TomWagg
Thanks very much for all of the comments/issues so far @dbkeitel, I really appreciate all of the improvements you're helping to make to the package and its documentation! ☺️ And I'm glad to hear you like the tests and documentation, thanks! 😁

> Another comment, mostly thinking towards future maintainability, is that there are a bunch of things hardcoded to particular models from particular papers (e.g. LISA instrumental and confusion noise as well as the set of verification binaries), which may need to change in the future. However it seems to me that the architecture of the package is laid out well enough that this should be possible even in a backwards-compatible way, by e.g. introducing new string arguments for the names of models into those functions.

This is definitely a concern I share! There's no doubt that there will be new models for sensitivity curve/confusion noise etc as we approach the launch of these sorts of space-based detectors so some of the hardcoded stuff will become obsolete. My 'plan' for this is that we will continue to implement new psd `instrument`s and confusion noise `model`s that can be chosen when computing SNRs etc. I imagine that in future we'll make `lisa_psd` or `tianqin_psd` aliases for whatever the latest models are (but also allow people to run things like `robson19_psd` instead).

So in an ideal world future users who have created a new model for something could create issues (or if we're lucky, pull requests) with the functions that are used and then we can add them in. And in the meantime they can always use a custom instrument in `power_spectral_density` (and even custom confusion noise once [PR 72](https://github.com/TeamLEGWORK/LEGWORK/pull/72) is merged).
