### 2022-01-07 #3998 @dbkeitel
Yes, direct aliasing is a little different from the wrapper-function-with-str-argument approach I was thinking of, but it sounds like a good and equivalent future model too! Whenever the default changes, you'll just have to be very explicit in your changelogs.
