### 2022-01-07 TeamLEGWORK/LEGWORK#79 @dbkeitel
some naive questions on PN, spins, harmonics, and mergers
Hi @TomWagg @katiebreivik while doing the JOSS review, but sort of beyond its scope, I came up with some very naive/basic questions about the assumptions and range of validity for what you've chosen to implement. It's not at all that I doubt any of it is appropriate for the intended LISA-and-friends use cases, just that I didn't find it really spelled out explicitly in the documentation nor the longer ApJ paper. Not sure if this issue tracker is really the right place to ask, but if I didn't overlook something obvious, then probably the questions can lead to improving the documentation. So...

- You obviously comment that you use post-Newtonian approximations, but without digging into the classic references or the code itself, it's hard to figure out to which order, and what the corresponding limitations are.
- Specifically, I couldn't find any mention anywhere of compact object spins, and whether they would have any effect on evolution timescales, strains or SNRs within reasonable tolerance requirements.
- For strains/SNRs, it is said explicitly in various places that higher harmonics are only needed for eccentric orbits. But in principle, the full waveforms for quasi-circular inspirals also include (weaker) higher modes. It's probably very true that they can be neglected until close to merger, but wouldn't this be worth a mention too?
- And for those binaries that do merge within the detector band, is there any rule of thumb that can be given for how long the strains stay accurate and how much the SNR may be off?
