### 2022-01-11 af4f179
remove "auto" option from interpolate_g
