### 2022-01-12 eb72910
change warning formatting so there aren't gaps
