### 2022-01-12 d150d2b
add warnings if timesteps are past the merger
