### 2022-01-12 aeadaee
make dimensions work properly with t_merge
