### 2022-01-12 TeamLEGWORK/LEGWORK#79 @katiebreivik
Thanks for the questions @dbkeitel! The LEGWORK team is not particularly well versed in GR theory (we come from the progenitor side), so I hope that my answers make sense! Please do let me know if anything I say rings any alarm bells!

> * You obviously comment that you use post-Newtonian approximations, but without digging into the classic references or the code itself, it's hard to figure out to which order, and what the corresponding limitations are.

What we should say (and will add to the documentation/paper) is "lowest-order post-Newtonian approximations". We don't include anything more than the energy radiated in GWs which is carried away from the binary. This means that things like spin-orbit coupling or radiation reaction are completely unaccounted for.

> * Specifically, I couldn't find any mention anywhere of compact object spins, and whether they would have any effect on evolution timescales, strains or SNRs within reasonable tolerance requirements.

The Peters evolution equations don't take spins into account at all. This is a very solid assumption for white dwarfs which should broadly be completely (or near-completely) tidally locked and aligned. For NSs and low-mass BHs, the spin-orbit coupling is also expected to be basically negligible. But, the place where our calculation very likely breaks down is for the higher-mass BHs (think M ~ 30). I agree that this warrants a warning in the docs and paper!

> * For strains/SNRs, it is said explicitly in various places that higher harmonics are only needed for eccentric orbits. But in principle, the full waveforms for quasi-circular inspirals also include (weaker) higher modes. It's probably very true that they can be neglected until close to merger, but wouldn't this be worth a mention too?

Still working on this one! My hunch is that it is very unlikely that the higher modes will affect the SNR too much for circular binaries at mHz frequencies, but I'm still digging for references to back that up!

> * And for those binaries that do merge within the detector band, is there any rule of thumb that can be given for how long the strains stay accurate and how much the SNR may be off?

The only stellar-mass systems which could merge in the LISA band are stripped stars (helium cores) orbiting a NS or WD. In this case, the SNR for the merger is very likely to be wildly inaccurate because gas dynamics will probably take over any GW effects. These are very rare sources though (i.e. [Götberg+2020](https://ui.adsabs.harvard.edu/abs/2020ApJ...904...56G/abstract)). Otherwise, the LISA frequency band is still too low to observe any stellar-mass objects actually making it to merger. This goes back to the point above for systems where spin-orbit coupling is strong (i.e. high mass BBH systems with non-aligned spins); in this case, the SNR is probably not totally accurate.
