### 2022-01-12 TeamLEGWORK/LEGWORK#79 @dbkeitel
Thanks for the detailed reply @katiebreivik ! I'd say we're on the same page then: all these approximations seem perfectly justified in the targeted use cases, and just need to be called out here and there in the documentation.

> The LEGWORK team is not particularly well versed in GR theory

Me neither, I'm much more of a hands-on analyst than a real theorist too.

And btw the stripped cores sound fascinating, I wasn't really aware of those as a source class yet!
