### 2022-01-13 e58c787
adding in discussion of limitations and scope and changelog update
