### 2022-01-13 TeamLEGWORK/LEGWORK#79 @katiebreivik
Just to tie up the last loose end on higher order modes: Since we are using calculations based on Peters 1963 and Peters & Mathews 1964, we are basically operating at 0 PN order which means that we only consider the dominant l=2, m=2 modes and so for circular or quasicircular binaries, the n=2 harmonic is the *only* harmonic for GW emission.
