### 2022-01-13 bfee5b1
Minor formatting issue in changelog
