+++ paper.bib
@@ -867,3 +867,38 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvD.102f3021H},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Littenberg+2020,
+       author = {{Littenberg}, Tyson B. and {Cornish}, Neil J. and {Lackeos}, Kristen and {Robson}, Travis},
+        title = "{Global analysis of the gravitational wave signal from Galactic binaries}",
+      journal = {\prd},
+     keywords = {General Relativity and Quantum Cosmology, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2020,
+        month = jun,
+       volume = {101},
+       number = {12},
+          eid = {123021},
+        pages = {123021},
+          doi = {10.1103/PhysRevD.101.123021},
+archivePrefix = {arXiv},
+       eprint = {2004.08464},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvD.101l3021L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Yi+2021,
+       author = {{Yi}, Shu-Xu and {Nelemans}, Gijs and {Brinkerink}, Christiaan and {Kostrzewa-Rutkowska}, Zuzanna and {Timmer}, Sjoerd T. and {Stoppa}, Fiorenzo and {Rossi}, Elena M. and {Portegies Zwart}, Simon F.},
+        title = "{The Gravitational Wave Universe Toolbox: A software package to simulate observation of the Gravitational Wave Universe with different detectors}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2021,
+        month = jun,
+          eid = {arXiv:2106.13662},
+        pages = {arXiv:2106.13662},
+archivePrefix = {arXiv},
+       eprint = {2106.13662},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021arXiv210613662Y},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
