+++ paper.bib
@@ -886,19 +886,3 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvD.101l3021L},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-@ARTICLE{Yi+2021,
-       author = {{Yi}, Shu-Xu and {Nelemans}, Gijs and {Brinkerink}, Christiaan and {Kostrzewa-Rutkowska}, Zuzanna and {Timmer}, Sjoerd T. and {Stoppa}, Fiorenzo and {Rossi}, Elena M. and {Portegies Zwart}, Simon F.},
-        title = "{The Gravitational Wave Universe Toolbox: A software package to simulate observation of the Gravitational Wave Universe with different detectors}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2021,
-        month = jun,
-          eid = {arXiv:2106.13662},
-        pages = {arXiv:2106.13662},
-archivePrefix = {arXiv},
-       eprint = {2106.13662},
- primaryClass = {astro-ph.HE},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021arXiv210613662Y},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
