### 2022-01-14 71387e8
Update paper.md
Co-authored-by: Tom Wagg <tomjwagg@gmail.com>
