### 2022-01-14 #3998 @katiebreivik
> @katiebreivik I did a few quick checks with `lalsimulation.SimInspiralChooseTDWaveform`. Just naively going to very low starting frequencies at fixed `f_min` blows up runtime and memory use, predictably. Reducing time resolution to make up for that works to some degree, but seems to run into some weird indexing issues when reducing it too much. If you're interested in knowing the transition region where it stops making sense, I can ask some of my more knowledgeable colleagues next week; but forgetting about this tangent and just citing some actually LISA-focused tools would be perfectly fine for me as well.

Thanks for checking on this @dbkeitel! We did decide decide to go with LISA-focused tools only.

With the recently merged [PR](https://github.com/TeamLEGWORK/LEGWORK/pull/83) on our joss branch, I think we have covered your concerns/questions/suggestions so far including updates to documentation and the paper. Please don't hesitate to point out anything we may have missed though!
