### 2022-01-14 TeamLEGWORK/LEGWORK#84 @cmbiwer
Numpy version for correct installation
I am following the instructions and there appears to be a problem with the version of numpy perhaps. I get this error trying to import anything from legwork:
```
>>> import legwork.source as source
RuntimeError: module compiled against API version 0xe but this version of numpy is 0xd
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/legwork/__init__.py", line 1, in <module>
    from . import evol, psd, snr, source, strain, utils, visualisation
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/legwork/evol.py", line 4, in <module>
    import legwork.utils as utils
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/legwork/utils.py", line 4, in <module>
    from astropy import constants as c
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/__init__.py", line 42, in <module>
    from . import config as _config  # noqa: E402
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/config/__init__.py", line 10, in <module>
    from .configuration import *
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/config/configuration.py", line 24, in <module>
    from astropy.utils import find_current_module, silence
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/utils/__init__.py", line 17, in <module>
    from .codegen import *  # noqa
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/utils/codegen.py", line 13, in <module>
    from .introspection import find_current_module
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/utils/introspection.py", line 14, in <module>
    from astropy.utils.decorators import deprecated_renamed_argument
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/utils/decorators.py", line 15, in <module>
    from .exceptions import (AstropyDeprecationWarning, AstropyUserWarning,
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/astropy/utils/exceptions.py", line 11, in <module>
    from erfa import ErfaError, ErfaWarning  # noqa
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/erfa/__init__.py", line 5, in <module>
    from .version import version as __version__  # noqa
  File "/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/erfa/version.py", line 25, in <module>
    from . import ufunc
ImportError: numpy.core.multiarray failed to import
```

I would assume the numpy version is incorrect. The ``requirements.txt`` lists numpy between 1.16 and 1.20. When I ran the ``pip install legwork`` it installed 1.19.5. My ``conda list`` is:
```
# packages in environment at /Users/cmbiwer/anaconda3/envs/legwork:
#
# Name                    Version                   Build  Channel
appnope                   0.1.2           py39hecd8cb5_1001
argon2-cffi               20.1.0           py39h9ed2024_1
astropy                   5.0              py39h7588534_0
async_generator           1.10               pyhd3eb1b0_0
attrs                     21.2.0             pyhd3eb1b0_0
backcall                  0.2.0              pyhd3eb1b0_0
blas                      1.0                         mkl
bleach                    4.1.0              pyhd3eb1b0_0
bottleneck                1.3.2            py39he3068b8_1
brotli                    1.0.9                hb1e8313_2
ca-certificates           2021.10.26           hecd8cb5_2
certifi                   2021.10.8        py39hecd8cb5_2
cffi                      1.14.6           py39h2125817_0
cycler                    0.11.0             pyhd3eb1b0_0
debugpy                   1.5.1            py39he9d5cce_0
decorator                 5.1.0              pyhd3eb1b0_0
defusedxml                0.7.1              pyhd3eb1b0_0
entrypoints               0.3              py39hecd8cb5_0
fonttools                 4.25.0             pyhd3eb1b0_0
freetype                  2.11.0               hd8bbffd_0
giflib                    5.2.1                haf1e3a3_0
icu                       58.2                 h0a44026_3
importlib-metadata        4.8.2            py39hecd8cb5_0
importlib_metadata        4.8.2                hd3eb1b0_0
intel-openmp              2021.4.0          hecd8cb5_3538
ipykernel                 6.4.1            py39hecd8cb5_1
ipython                   7.29.0           py39h01d92e1_0
ipython_genutils          0.2.0              pyhd3eb1b0_1
ipywidgets                7.6.5              pyhd3eb1b0_1
jedi                      0.18.0           py39hecd8cb5_1
jinja2                    3.0.2              pyhd3eb1b0_0
jpeg                      9d                   h9ed2024_0
jsonschema                3.2.0              pyhd3eb1b0_2
jupyter                   1.0.0            py39hecd8cb5_7
jupyter_client            7.1.0              pyhd3eb1b0_0
jupyter_console           6.4.0              pyhd3eb1b0_0
jupyter_core              4.9.1            py39hecd8cb5_0
jupyterlab_pygments       0.1.2                      py_0
jupyterlab_widgets        1.0.0              pyhd3eb1b0_1
kiwisolver                1.3.1            py39h23ab428_0
lcms2                     2.12                 hf1fd2bf_0
legwork                   0.2.1                    pypi_0    pypi
libcxx                    12.0.0               h2f01273_0
libffi                    3.3                  hb1e8313_2
libgfortran               3.0.1                h93005f0_2
libllvm11                 11.1.0               h9b2ccf5_0
libpng                    1.6.37               ha441bb4_0
libsodium                 1.0.18               h1de35cc_0
libtiff                   4.2.0                h87d7836_0
libwebp                   1.2.0                hacca55c_0
libwebp-base              1.2.0                h9ed2024_0
llvm-openmp               12.0.0               h0dcd299_1
llvmlite                  0.36.0                   pypi_0    pypi
lz4-c                     1.9.3                h23ab428_1
markupsafe                2.0.1            py39h9ed2024_0
matplotlib                3.5.0            py39hecd8cb5_0
matplotlib-base           3.5.0            py39h4f681db_0
matplotlib-inline         0.1.2              pyhd3eb1b0_2
mistune                   0.8.4           py39h9ed2024_1000
mkl                       2021.4.0           hecd8cb5_637
mkl-service               2.4.0            py39h9ed2024_0
mkl_fft                   1.3.1            py39h4ab4a9b_0
mkl_random                1.2.2            py39hb2f4e1b_0
munkres                   1.1.4                      py_0
nbclient                  0.5.3              pyhd3eb1b0_0
nbconvert                 6.3.0            py39hecd8cb5_0
nbformat                  5.1.3              pyhd3eb1b0_0
ncurses                   6.3                  hca72f7f_2
nest-asyncio              1.5.1              pyhd3eb1b0_0
notebook                  6.4.6            py39hecd8cb5_0
numba                     0.53.1                   pypi_0    pypi
numexpr                   2.8.1            py39h2e5f0a9_0
numpy                     1.19.5                   pypi_0    pypi
olefile                   0.46               pyhd3eb1b0_0
openssl                   1.1.1m               hca72f7f_0
packaging                 21.3               pyhd3eb1b0_0
pandas                    1.3.5            py39h743cdd8_0
pandocfilters             1.4.3            py39hecd8cb5_1
parso                     0.8.3              pyhd3eb1b0_0
pexpect                   4.8.0              pyhd3eb1b0_3
pickleshare               0.7.5           pyhd3eb1b0_1003
pillow                    8.4.0            py39h98e4679_0
pip                       21.2.4           py39hecd8cb5_0
prometheus_client         0.12.0             pyhd3eb1b0_0
prompt-toolkit            3.0.20             pyhd3eb1b0_0
prompt_toolkit            3.0.20               hd3eb1b0_0
ptyprocess                0.7.0              pyhd3eb1b0_2
pycparser                 2.21               pyhd3eb1b0_0
pyerfa                    2.0.0            py39h9ed2024_0
pygments                  2.10.0             pyhd3eb1b0_0
pyparsing                 3.0.4              pyhd3eb1b0_0
pyqt                      5.9.2            py39h23ab428_6
pyrsistent                0.18.0           py39hca72f7f_0
python                    3.9.7                h88f2d9e_1
python-dateutil           2.8.2              pyhd3eb1b0_0
pytz                      2021.3             pyhd3eb1b0_0
pyyaml                    6.0              py39hca72f7f_1
pyzmq                     22.3.0           py39he9d5cce_2
qt                        5.9.7                h468cd18_1
qtconsole                 5.1.1              pyhd3eb1b0_0
qtpy                      1.10.0             pyhd3eb1b0_0
readline                  8.1.2                hca72f7f_1
schwimmbad                0.3.2                    pypi_0    pypi
scipy                     1.7.3            py39h8c7af03_0
seaborn                   0.11.2             pyhd3eb1b0_0
send2trash                1.8.0              pyhd3eb1b0_1
setuptools                58.0.4           py39hecd8cb5_0
sip                       4.19.13          py39h23ab428_0
six                       1.16.0             pyhd3eb1b0_0
sqlite                    3.37.0               h707629a_0
terminado                 0.9.4            py39hecd8cb5_0
testpath                  0.5.0              pyhd3eb1b0_0
tk                        8.6.11               h7bc2e8c_0
tornado                   6.1              py39h9ed2024_0
traitlets                 5.1.1              pyhd3eb1b0_0
tzdata                    2021e                hda174b7_0
wcwidth                   0.2.5              pyhd3eb1b0_0
webencodings              0.5.1            py39hecd8cb5_1
wheel                     0.37.1             pyhd3eb1b0_0
widgetsnbextension        3.5.1            py39hecd8cb5_0
xz                        5.2.5                h1de35cc_0
yaml                      0.2.5                haf1e3a3_0
zeromq                    4.3.4                h23ab428_0
zipp                      3.7.0              pyhd3eb1b0_0
zlib                      1.2.11               h4dc903c_4
zstd                      1.4.9                h322a384_0
```
And the numpy is in the env:
```
>>> import numpy
>>> numpy
<module 'numpy' from '/Users/cmbiwer/anaconda3/envs/legwork/lib/python3.9/site-packages/numpy/__init__.py'>
>>> numpy.__version__
'1.19.5'
```
