### 2022-01-14 TeamLEGWORK/LEGWORK#84 @cmbiwer
Upgrading to numpy 1.22.1 which is out of your spec seems to load the library. Though the initial import takes a long time for the size of this library (~10 seconds) until it completes.
