### 2022-01-14 #3998 @cmbiwer
I've posted an issue here: https://github.com/TeamLEGWORK/LEGWORK/issues/84

Following the instructions on the installation page:
```
conda create --name legwork numpy scipy astropy numba matplotlib seaborn jupyter ipython
conda activate legwork
pip install schwimmbad legwork
```
Is not getting me a working package.
