### 2022-01-14 #3998 @cmbiwer
As far as the paper, I think it is well written and describes the purpose. The only comment I have on the paper is the state of the field is summarized as "So far, most studies made use of custom made codes which have not been made
47 publicly available." I wouldn't ask the authors to give examples of closed-source code. But do the authors know of publicly available codes that could be cited? The use of the word "some" implies to me that they could give examples.
