### 2022-01-18 56cb0a9
removing upper bounds on numpy and numba versions based on new numba release
