+++ requirements.txt
@@ -1,5 +1,5 @@
-numba >= 0.50,< 0.54'
-numpy >= 1.16,< 1.20'
+numba >= 0.50'
+numpy >= 1.16,'
 astropy >= 4.0'
 scipy >= 1.5.0'
 matplotlib >= 3.3.2'
