### 2022-01-18 #3998 @katiebreivik
Thanks @cmbiwer for pointing out the install problems based on the numpy version. We've just merged in a [PR](https://github.com/TeamLEGWORK/LEGWORK/pull/85) which fixes this issue, as well as removes the annoying package downgrades which happened because of the upper bounds on numba and numpy.

With regard to the paper and citation of tools. We've updated the draft based on @dbkeitel's suggestions above and I think the updated draft provides the citations you were after. I'll try to recompile the paper here now (though I'm not exactly sure whether this will work!).
