### 2022-01-24 TeamLEGWORK/LEGWORK#75 @cmbiwer
I am still seeing this error in the quick start tutorial. From LEGWORK main branch githash: 7e3d98b

```
Traceback (most recent call last):
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 747, in to_value
    scale = self.unit._to(unit)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1097, in _to
    f"'{self!r}' is not a scaled version of '{other!r}'")
astropy.units.core.UnitConversionError: 'Unit("Hz")' is not a scaled version of 'Unit(dimensionless)'

Traceback (most recent call last):
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 1090, in __float__
    return float(self.to_value(dimensionless_unscaled))
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 750, in to_value
    value = self._to_value(unit, equivalencies)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 670, in _to_value
    equivalencies=equivalencies)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1134, in to
    equivalencies=equivalencies)(value)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1064, in _get_converter
    raise exc
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1050, in _get_converter
    self, other, self._normalize_equivalencies(equivalencies))
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1026, in _apply_equivalencies
    f"{unit_str} and {other_str} are not convertible")
astropy.units.core.UnitConversionError: 'Hz' (frequency) and '' (dimensionless) are not convertible

Traceback (most recent call last):
  File "test.py", line 35, in <module>
    vis.plot_sensitivity_curve(frequency_range=frequency_range, fig=fig, ax=right_ax)
  File "/Users/cmbiwer/src/LEGWORK/legwork/visualisation.py", line 314, in plot_sensitivity_curve
    ax.fill_between(frequency_range, 0 * noise_amplitude.unit, noise_amplitude, alpha=alpha, color=color)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/__init__.py", line 1412, in inner
    return func(ax, *map(sanitize_sequence, args), **kwargs)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/axes/_axes.py", line 5254, in fill_between
    where=where, interpolate=interpolate, step=step, **kwargs)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/axes/_axes.py", line 5225, in _fill_between_x_or_y
    pts[0] = start
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 1092, in __float__
    raise TypeError('only dimensionless scalar quantities can be '
TypeError: only dimensionless scalar quantities can be converted to Python scalars
```
