### 2022-01-24 TeamLEGWORK/LEGWORK#75 @cmbiwer
I also see this error for ``fig, ax = sources.plot_sources_on_sc(snr_cutoff=7, disttype="kde")`` at: https://legwork.readthedocs.io/en/latest/notebooks/Source.html
