### 2022-01-24 TeamLEGWORK/LEGWORK#75 @cmbiwer
Encountered another units error with
```# plot the default as an area
ax.fill_between(e_range, 0, t_merge, label="Default", alpha=0.2)
```
On https://legwork.readthedocs.io/en/latest/notebooks/Evolution.html

```
Traceback (most recent call last):
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 747, in to_value
    scale = self.unit._to(unit)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1097, in _to
    f"'{self!r}' is not a scaled version of '{other!r}'")
astropy.units.core.UnitConversionError: 'Unit("Gyr")' is not a scaled version of 'Unit(dimensionless)'

Traceback (most recent call last):
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 1090, in __float__
    return float(self.to_value(dimensionless_unscaled))
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 750, in to_value
    value = self._to_value(unit, equivalencies)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 670, in _to_value
    equivalencies=equivalencies)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1134, in to
    equivalencies=equivalencies)(value)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1064, in _get_converter
    raise exc
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1050, in _get_converter
    self, other, self._normalize_equivalencies(equivalencies))
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/core.py", line 1026, in _apply_equivalencies
    f"{unit_str} and {other_str} are not convertible")
astropy.units.core.UnitConversionError: 'Gyr' (time) and '' (dimensionless) are not convertible

Traceback (most recent call last):
  File "test4.py", line 43, in <module>
    ax.fill_between(e_range, 0, t_merge, label="Default", alpha=0.2)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/__init__.py", line 1412, in inner
    return func(ax, *map(sanitize_sequence, args), **kwargs)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/axes/_axes.py", line 5254, in fill_between
    where=where, interpolate=interpolate, step=step, **kwargs)
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/matplotlib/axes/_axes.py", line 5225, in _fill_between_x_or_y
    pts[0] = start
  File "/Users/cmbiwer/anaconda3/envs/legwork2/lib/python3.7/site-packages/astropy/units/quantity.py", line 1092, in __float__
    raise TypeError('only dimensionless scalar quantities can be '
TypeError: only dimensionless scalar quantities can be converted to Python scalars
```
