### 2022-01-24 TeamLEGWORK/LEGWORK#86 @cmbiwer
Horizon distance demo missing variable initialization
**Which documentation page is causing an issue?**
https://legwork.readthedocs.io/en/latest/demos/HorizonDistance.html

**What is the problem?**
The variable for font size (``fs``) isn't defined on this page.

**Suggested solution (optional)**
Define ``fs`` to be some number.
