### 2022-01-24 TeamLEGWORK/LEGWORK#86 @cmbiwer
Also this page: https://legwork.readthedocs.io/en/latest/demos/CompareSensitivityCurves.html
