### 2022-01-24 #3998 @cmbiwer
I've gone through the package again after the installation was updated. The package is well documented and the automated tests have good coverage. I marked off nearly every item on the checklist. However, the only functionality I have found not working on the ``main`` branch is still the unit error using ``plot_sensitivity_curve`` and also ``ax.fill_between`` that was reported here: https://github.com/TeamLEGWORK/LEGWORK/issues/75

There was also a minor issue in the documentation pages for the demos, the variable ``fs`` seems to be missing from a number of those pages which I made an issue for here: https://github.com/TeamLEGWORK/LEGWORK/issues/86

If the authors could resolve this plotting issue with ``plot_sensitivity_curve`` and ``ax.fill_between``, then I would say LEGWORK meets JOSS's requirements for publication, and I would recommend its publication.
