### 2022-01-24 TeamLEGWORK/LEGWORK#86 @TomWagg
Thanks for the issue! So the Jupyter notebooks from which the demos are generated _do_ have a definition of `fs`. See [here](https://github.com/TeamLEGWORK/LEGWORK/blob/main/docs/demos/HorizonDistance.ipynb) for example.

I was planning that if people wanted to actually reproduced the demos they would use the link to the notebook at the top of the page. I hid a code block that looks like this
```

%config InlineBackend.figure_format = 'retina'

plt.rc('font', family='serif')
plt.rcParams['text.usetex'] = False
fs = 24

# update various fontsizes to match
params = {'figure.figsize': (12, 8),
          'legend.fontsize': fs,
          'axes.labelsize': fs,
          'xtick.labelsize': 0.9 * fs,
          'ytick.labelsize': 0.9 * fs,
          'axes.linewidth': 1.1,
          'xtick.major.size': 7,
          'xtick.minor.size': 4,
          'ytick.major.size': 7,
          'ytick.minor.size': 4}
plt.rcParams.update(params)
```
 in each of the demos as I thought it used a fair bit of space but didn't really add anything to the demo.

Do you think it would be more useful to include this code block in the demo?
