### 2022-01-24 c4707f0
remove units if it is a histogram
