+++ legwork/source.py
@@ -1041,8 +1041,7 @@ class Source():
 
         # plot it!
         if ystr is not None:
-            return vis.plot_2D_dist(x=x[which_sources], y=y[which_sources],
-                                    weights=weights, **kwargs)
+            return vis.plot_2D_dist(x=x[which_sources], y=y[which_sources], weights=weights, **kwargs)
         else:
             return vis.plot_1D_dist(x=x[which_sources], weights=weights, **kwargs)
 
