### 2022-01-24 TeamLEGWORK/LEGWORK#75 @TomWagg
Thanks for pointing this out @cmbiwer, it seems `ax.fill_between` has become less tolerant of stray units. I've fixed this by adding `quantity_support()` in the relevant locations, so version 0.2.3 should work fine again.
