### 2022-01-25 TeamLEGWORK/LEGWORK#86 @cmbiwer
``jupyter`` does not seem to be in the LEGWORK environment.

I'd suggest you either include ``jupyter`` in the environment and give instructions on how you want the user to execute the tutorials through the notebook. Or include everything you need to reproduce the images and results on the webpage the user is following.
