### 2022-01-25 TeamLEGWORK/LEGWORK#86 @TomWagg
Gotcha yeah as you say it's there for the stable case but not the developer. You do bring something to my attention though: creating the conda environment with all of those packages may not grab the right versions and then they may need to be re-installed during the LEGWORK installation.

I think perhaps this could be a better solution:
```
    conda create --name legwork
    conda activate legwork

    pip install legwork

    conda install jupyter ipython
```
where the installation of LEGWORK will automatically install the right dependencies using [requirements.txt](https://github.com/TeamLEGWORK/LEGWORK/blob/main/requirements.txt) and then the last line is optional if you also want to use Jupyter notebooks for the demos and notebooks.

For the development page I would change the third line to
```
    git clone https://github.com/TeamLEGWORK/LEGWORK
    cd LEGWORK
    pip install .
```

Would this be a reasonable solution for you? If so I can update the installation page right away.
