### 2022-01-25 TeamLEGWORK/LEGWORK#86 @cmbiwer
> Gotcha yeah as you say it's there for the stable case but not the developer. You do bring something to my attention though: creating the conda environment with all of those packages may not grab the right versions and then they may need to be re-installed during the LEGWORK installation.
>
> I think perhaps this could be a better solution:
>
> ```
>     conda create --name legwork
>     conda activate legwork
>
>     pip install legwork
>
>     conda install jupyter ipython
> ```

I think you should do this. This was one of my comments that what's there now is somewhat inefficient: https://github.com/openjournals/joss-reviews/issues/3998#issuecomment-1013544655

> Would this be a reasonable solution for you? If so I can update the installation page right away.

I would say the installation and documentation are two different issues. On the documentation page, I think you should have one of those notes/warning boxes (i.e. use ``.. warning:`` or ``.. note:`` https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-warning) in Sphinx that the webpage only contains portions of the notebook and the user should run the notebook by doing x, y, z commands; and show what those commands are. If that's what you want the user to do.

> Also perhaps it would be useful to note on the main demos page that there are also .py files that recreate all of the demos in the LEGWORK-paper repository if people don't want to use jupyter notebooks? (here: https://github.com/TeamLEGWORK/LEGWORK-paper/tree/main/src/figures)

I think its enough to just state how you want the user to run the demos. If you want to maintain the notebooks and have them use the notebooks, then its probably not worthwhile to link to these other scripts. Its just one extra thing you would have to remember maintain.
