### 2022-01-25 TeamLEGWORK/LEGWORK#86 @TomWagg
Ah my apologies for missing that comment! I'll definitely change this then!

Okay I see what you mean, I'll add a note box to the demos page explaining how they should be run! Yeah I don't mind maintaining the demos, I'll probably also try automating the testing of them though, will make a new issue for that.
