### 2022-01-25 4a69909
change install instructions, smarter re versions
