+++ changelog.md
@@ -106,3 +106,4 @@ Small changes to visualisation code and updates to tutorials/demos with the new 
 ## 0.2.3
 *TW 23/01/22*
 - [Issue [#75](https://github.com/TeamLEGWORK/LEGWORK/issues/75)] Fix mixing quantities with floats when plotting
+- [Issue [#86](https://github.com/TeamLEGWORK/LEGWORK/issues/86)] Clarify how notebooks should be run and update the installation instructions
