### 2022-01-26 #3998 @TomWagg
[PR 87](https://github.com/TeamLEGWORK/LEGWORK/pull/87) is merged into LEGWORK (bringing us up to v.0.2.3)! 😁 I _think_ this means we have now addressed each of @cmbiwer's comments (but please do let me know if not!). If that's the case I think we'll also need the final go-ahead from @dbkeitel before moving forward 🙂
