### 2022-01-27 #3998 @cmbiwer
I have check that PR#87 has addressed my last issues. And I have marked off all the items on the checklist now.

In my opinion, LEGWORK meets JOSS's requirements for publication and I recommend it is published in JOSS.
