### 2022-01-27 #3998 @dbkeitel
Hi @TomWagg the updated paper looks good, seems you just missed one of my citation requests: please cite the LIGO first detection paper where "the direct detection" is mentioned (now L39).
