### 2022-01-27 TeamLEGWORK/LEGWORK#90 @dbkeitel
confusion noise functions not showing up in docs
**Which documentation page is causing an issue?**
https://legwork.readthedocs.io/en/latest/modules.html#module-legwork.psd

**What is the problem?**
The LISA/TQ function documentations refer to `legwork.psd.get_confusion_noise()` but that identifier is not clickable, and I can't actually find that function (nor the specific implementations of confusion noise) anywhere in the API docs.

**Suggested solution (optional)**
The docstrings in https://github.com/TeamLEGWORK/LEGWORK/blob/main/legwork/psd.py look valid to me, so maybe the functions need to be added manually somewhere to sphinx...?
