### 2022-01-27 #3998 @dbkeitel
The package itself is also almost ready to sign off from my side, with most of my issues resolved well. A few open points though:
- [ ] the updated conda installation instructions don't work for me -> https://github.com/TeamLEGWORK/LEGWORK/issues/89
- [ ] another small documentation issue -> https://github.com/TeamLEGWORK/LEGWORK/issues/90
- [ ] I'm not entirely sure if any specific comments have been added to the documentation anywhere to address the (very enlightening!) discussions we had on the physical assumptions in https://github.com/TeamLEGWORK/LEGWORK/issues/79). I only skimmed the tutorials in this second round and maybe I missed them, but e.g. in https://legwork.readthedocs.io/en/latest/notebooks/Source.html when it is first said that needing anything beyond n=2 is equivalent to eccentricity, it would be nice to briefly point out the approximations this is valid under.

The third point is optional I'd say and up to the authors. Once the one citation is added to the paper and the installation instructions and API docs fixed, I'm happy to sign off the review.
