### 2022-01-27 TeamLEGWORK/LEGWORK#90 @TomWagg
Ah nice catch! Sphinx only makes docs for functions in the __all__ list at the top, so I just need to add them to that, will do!
