### 2022-01-27 TeamLEGWORK/LEGWORK#89 @TomWagg
Haha it seems every "fix" we do for these dependencies breaks it in a different way 😂

I think it probably makes sense to just include the version numbers manually on the instruction page since users would need to download that .yml file manually anyway right? Might as well just have a copy-able code block instead.

I'll add these back and add notes to the requirement.txt file to remember to always update in both locations.
