### 2022-01-27 TeamLEGWORK/LEGWORK#89 @dbkeitel
The advantage of providing a .yml is that at a later point, people can re-download it and do
```
conda env update --file environment.yml  --prune
```
to get all dependencies updated consistently.

But your suggestion is certainly the simplest.
