### 2022-01-27 #3998 @katiebreivik
Thanks @dbkeitel! We will get to work on the first two open points. For the third, we decided to add a scope/limitations section to the docs ([here](https://legwork.readthedocs.io/en/latest/limitations.html)). We settled on this because we didn't want to repeat the same explanation in many different sections of the docs. I'd be interested to hear your thoughts on whether this is appropriate. Maybe we should add a link to the tab in the Quickstart to make sure that folks are aware?
