### 2022-01-27 #3998 @dbkeitel
Thanks @katiebreivik I had completely overlooked that section! I agree it's a good solution. A link at the very end of the quickstart may indeed be a good idea too,
