### 2022-01-27 e8e8904
make sure confusion noise functions appear in docs
