### 2022-01-27 TeamLEGWORK/LEGWORK#89 @TomWagg
Okay I see what you mean. I don't know that much about package managers and whatnot but I think that if, at some point in the future, a user wants to update legwork they could run
```
pip install legwork --upgrade
```

My question is then, when that upgrade occurs, would it also check whether the dependencies could be upgraded without needing to update the environment with the new .yml file?
