### 2022-01-27 205ff0c
add specific versions to conda environment
