### 2022-01-28 TeamLEGWORK/LEGWORK#89 @TomWagg
Having googled this a bit I've realised that pip upgrades are completely invisible to conda so I do see how the environment.yml file could be very useful here.

I'm going to add it for the stable case but I _think_ there isn't much point with the development version since you always have to use pip to install it anyway.
