+++ environment.yml
@@ -1,0 +1,16 @@
+name: legwork
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python >= 3.7
+  - pip >= 21.0.0
+  - numba >= 0.50
+  - numpy >= 1.16
+  - astropy >= 4.0
+  - scipy >= 1.5.0
+  - matplotlib >= 3.3.2
+  - seaborn >= 0.11.1
+  - schwimmbad >= 0.3.2
+  - pip:
+    - legwork >= 0.2.4
