+++ setup.cfg
@@ -9,6 +9,11 @@ description = A python package to compute gravitational wave evolution and signa
 long_description = file: README.md
 long_description_content_type = text/markdown
 
+# NOTE: any changes in the requirements must also be reflected in
+#   - requirements.txt
+#   - environment.yml
+#   - install.rst
+
 [options]
 python_requires = >=3.7
 packages = find: 
