### 2022-01-28 28f37bd
update install instructions to use environment.yml
