### 2022-01-28 a60b799
add LIGO citation
