### 2022-01-28 TeamLEGWORK/LEGWORK#89 @dbkeitel
Trying to understand the differences between the stable and dev instructions, I've stumbled about non-intuitive (for me at least) behaviour regarding how conda isolates (or doesn't isolate) an env from system packages. Apparently this depends on whether the env's python version matches the system's:
- If I create an env with the same python version as my system, then...
  - the current stable instructions fail as described above
  - the current dev instructions sort-of-work, with dependencies not being actively installed into the env but available
- If I create an env with a different python version, then...
  - the current stable instructions work, with pip actually pulling in all dependencies and installing them into the env
  - the current dev instructions work the same

So unless you understand these things better than I do and know of a perfectly clean solution, I think changing the stable instructions to use a yml and leaving the dev case as it currently stands should at least give working installations in each situation.
