### 2022-01-28 TeamLEGWORK/LEGWORK#89 @dbkeitel
PS on your question regarding `pip install legwork --upgrade` my (also limited) understanding is that it may be more fragile than the yaml route.
