### 2022-01-28 0d82aed
let users install jupyter/ipython by themselves
