### 2022-01-28 #3998 @TomWagg
Hi again everyone, @katiebreivik and I have just merged in PR[#92](https://github.com/TeamLEGWORK/LEGWORK/pull/92) and [#93](https://github.com/TeamLEGWORK/LEGWORK/pull/93) and I believe they address all of @dbkeitel's remaining comments.

We've added the LIGO citation to the paper.md file, updated the installation instructions, added a link in the quickstart to the limitations page and made sure all of the functions are showing up in the documentation.

Let us know if there's anything else that we've missed! ☺️
