### 2022-01-31 #3998 @dbkeitel
@arfon I'm now happy to sign off on this review. The updated installation instructions work for me again and the paper looks good now.

@TomWagg just one minor bibliography detail, which in most journals the production department would resolve but here I guess falls on you: The GW150914 ref currently looks like this

![image](https://user-images.githubusercontent.com/28396587/151848579-c17793da-a627-4d8b-abaf-f59ecc4be7b7.png)

and should either have the `Virgo Collaboration` bit changed to `LIGO Scientific Collaboration and Virgo Collaboration` or removed entirely. (The bibstyle currently seems to treat it simply as the last entry in the human authors list, instead of as a separate collaboration element.)
