### 2022-01-31 #3998 @TomWagg
Wonderful, thanks @dbkeitel. I've fixed the citation (looks like the curly braces weren't quite right so it split LIGO and Virgo at the end).
