### 2022-02-01 #3998 @TomWagg
Exciting @arfon! Here is the information you asked for (and I should also thank @dbkeitel for explaining how to link GitHub to Zenodo in [an issue](https://github.com/TeamLEGWORK/LEGWORK/issues/63) a while ago!!)

- DOI: [10.5281/zenodo.5942553](https://doi.org/10.5281/zenodo.5821187)
- New release is here: https://github.com/TeamLEGWORK/LEGWORK/releases/tag/v0.3.0
- Zenodo archive is here: https://zenodo.org/record/5942553
