### 2022-03-16 #3998 @arfon
:wave: @TomWagg – looks like we have the AAS DOI now. I've opened a PR on the paper here https://github.com/TeamLEGWORK/LEGWORK/pull/95 but need you to verify that the journal name is correct.
