### 2022-03-17 #3998 @TomWagg
Thanks @arfon! I've updated the journal name to **Astrophysical Journal Supplement Series** and merged the PR!
