### 2022-03-17 #3998 @arfon
> Thanks @arfon! I've updated the journal name to Astrophysical Journal Supplement Series and merged the PR!

👍 thanks. The paper is updated now to include the crosslink to APJSS although it may take a few hours to show as updated on the JOSS site.
